+++ include/wx/pdfdoc_version.h
@@ -13,8 +13,8 @@
 
 #define PDFDOC_MAJOR_VERSION      1
 #define PDFDOC_MINOR_VERSION      2
-#define PDFDOC_RELEASE_NUMBER     0
+#define PDFDOC_RELEASE_NUMBER     1
 #define PDFDOC_SUBRELEASE_NUMBER  0
-#define PDFDOC_VERSION_STRING     "wxPdfDocument 1.2.0"
+#define PDFDOC_VERSION_STRING     "wxPdfDocument 1.2.1"
 
 #endif // _PDFDOC_VERSION_H_